+++ src/Repository/HousingGroupRepository.php
@@ -4,6 +4,7 @@ namespace App\Repository;
 
 use App\Entity\HousingGroup;
 use App\Entity\Lessor;
+use App\Model\SearchHousingGroupCriteriaModel;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
@@ -34,11 +35,35 @@ class HousingGroupRepository extends ServiceEntityRepository
         return array_column($results, 'city');
     }
 
-    public function getLessorHousingGroupListQueryBuilder(Lessor $lessor): QueryBuilder
+    public function getLessorDistinctCities(Lessor $lessor): array
     {
-        return $this
+        $results = $this->createQueryBuilder('h')
+            ->select('UPPER(h.address.city) as city')
+            ->distinct()
+            ->innerJoin('h.lessor', 'l', Join::WITH, 'l.id = :lessorId')
+            ->setParameter('lessorId', $lessor->getId())
+            ->orderBy('city', 'ASC')
+            ->getQuery()
+            ->getResult();
+
+        return array_column($results, 'city');
+    }
+
+    public function getLessorHousingGroupListQueryBuilder(Lessor $lessor, SearchHousingGroupCriteriaModel $searchHousingGroupCriteriaModel): QueryBuilder
+    {
+        $queryBuilder = $this
             ->createQueryBuilder('h')
                 ->innerJoin('h.lessor', 'l', Join::WITH, 'l.id = :lessorId')
                 ->setParameter('lessorId', $lessor->getId());
+
+        if ($searchHousingGroupCriteriaModel->getCity()) {
+            $queryBuilder
+                ->andWhere('UPPER(h.address.city) = :city')
+                ->setParameter('city', $searchHousingGroupCriteriaModel->getCity());
+        }
+
+        $queryBuilder = $queryBuilder->orderBy('h.name');
+
+        return $queryBuilder;
     }
 }